Rename classes
Make some changes in project by making names of classes and interfecases
more shortly and clearly in understending.
