+++ src/ss/ita/implementation/nik/NikSeven.java
@@ -1,8 +1,8 @@
-package ss.ita.kataimpl.liuba;
+package ss.ita.implementation.nik;
 
-import ss.ita.kata.KataSeven;
+import ss.ita.kata.Seven;
 
-public class LiubaKataSevenImpl implements KataSeven {
+public class NikSeven implements Seven {
     @Override
     public long newAvg(double[] arr, double navg) {
         return 0;
